// src/models/IoK8sApiAuthenticationV1beta1SelfSubjectReview/types.ts
import { IoK8sApiAuthenticationV1beta1SelfSubjectReviewStatusModelOptions as IoK8sApiAuthenticationV1beta1SelfSubjectReviewStatus } from '../IoK8sApiAuthenticationV1beta1SelfSubjectReviewStatus';
import { IoK8sApimachineryPkgApisMetaV1ObjectMetaModelOptions as IoK8sApimachineryPkgApisMetaV1ObjectMeta } from '../IoK8sApimachineryPkgApisMetaV1ObjectMeta';

export interface IoK8sApiAuthenticationV1beta1SelfSubjectReviewModelOptions {
  apiVersion?: string;
  kind?: string;
  metadata?: IoK8sApimachineryPkgApisMetaV1ObjectMeta;
  status?: IoK8sApiAuthenticationV1beta1SelfSubjectReviewStatus;
}